+++ apps/form-service/src/form/router/form.spec.ts
@@ -17,8 +17,6 @@ import {
   mapFormForSubmission,
   updateFormSubmissionDisposition,
   validateCriteria,
-  findDataRegisters,
-  createDataRegister,
 } from './form';
 import { getFormDefinitions } from './definition';
 
@@ -2461,433 +2459,6 @@ describe('form router', () => {
         params: { formId: 'test-form' },
       };
       expect(() => validateCriteria(req.query.criteria)).toThrow(InvalidOperationError);
-    });
-  });
-
-  describe('findDataRegisters', () => {
-    it('can create handler', () => {
-      const handler = findDataRegisters(directoryMock, tokenProviderMock);
-      expect(handler).toBeTruthy();
-    });
-
-    it('can find data registers and map descriptions from configuration definitions', async () => {
-      const req = {
-        tenant: { id: tenantId },
-      };
-
-      const res = {
-        status: jest.fn().mockReturnThis(),
-        send: jest.fn(),
-      };
-
-      const next = jest.fn();
-
-      directoryMock.getServiceUrl.mockResolvedValueOnce(new URL('https://configuration-service/configuration/v2'));
-
-      axiosMock.get
-        .mockResolvedValueOnce({
-          data: {
-            latest: {
-              configuration: {
-                'data-register:abcd': {
-                  description: '',
-                },
-                'data-register:fghj': {
-                  description: 'asdf',
-                },
-                'data-register:hello': {
-                  description: 'world',
-                },
-              },
-            },
-          },
-        })
-        .mockResolvedValueOnce({
-          data: {
-            results: [
-              {
-                namespace: 'data-register',
-                name: 'abcd',
-                latest: {
-                  configuration: ['123'],
-                },
-              },
-              {
-                namespace: 'data-register',
-                name: 'fghj',
-                latest: {
-                  configuration: ['123', '234', '3434', 'hellow woorld 123'],
-                },
-              },
-              {
-                namespace: 'data-register',
-                name: 'hello',
-                latest: {
-                  configuration: ['abc', '123', 'adf'],
-                },
-              },
-            ],
-          },
-        });
-
-      const handler = findDataRegisters(directoryMock, tokenProviderMock);
-      await handler(req as unknown as Request, res as unknown as Response, next);
-
-      expect(directoryMock.getServiceUrl.mock.calls[0][0].toString()).toBe('urn:ads:platform:configuration-service:v2');
-
-      expect(tokenProviderMock.getAccessToken).toHaveBeenCalled();
-      expect(axiosMock.get).toHaveBeenCalledTimes(2);
-
-      expect(axiosMock.get.mock.calls[0][0]).toContain('/configuration/platform/configuration-service');
-
-      expect(axiosMock.get.mock.calls[1][0]).toContain('/configuration/data-register');
-
-      expect(axiosMock.get.mock.calls[0][1]).toEqual(
-        expect.objectContaining({
-          headers: { Authorization: 'Bearer token' },
-          params: { tenantId: tenantId.toString() },
-        }),
-      );
-
-      expect(axiosMock.get.mock.calls[1][1]).toEqual(
-        expect.objectContaining({
-          headers: { Authorization: 'Bearer token' },
-          params: { tenantId: tenantId.toString() },
-          validateStatus: expect.any(Function),
-        }),
-      );
-
-      expect(res.status).toHaveBeenCalledWith(200);
-      expect(res.send).toHaveBeenCalledWith([
-        {
-          name: 'abcd',
-          namespace: 'data-register',
-          description: '',
-          entries: ['123'],
-        },
-        {
-          name: 'fghj',
-          namespace: 'data-register',
-          description: 'asdf',
-          entries: ['123', '234', '3434', 'hellow woorld 123'],
-        },
-        {
-          name: 'hello',
-          namespace: 'data-register',
-          description: 'world',
-          entries: ['abc', '123', 'adf'],
-        },
-      ]);
-
-      expect(next).not.toHaveBeenCalled();
-    });
-
-    it('uses empty description when register definition is missing', async () => {
-      const req = {
-        tenant: { id: tenantId },
-      };
-
-      const res = {
-        status: jest.fn().mockReturnThis(),
-        send: jest.fn(),
-      };
-
-      const next = jest.fn();
-
-      directoryMock.getServiceUrl.mockResolvedValueOnce(new URL('https://configuration-service/configuration/v2'));
-
-      axiosMock.get
-        .mockResolvedValueOnce({
-          data: {
-            latest: {
-              configuration: {},
-            },
-          },
-        })
-        .mockResolvedValueOnce({
-          data: {
-            results: [
-              {
-                namespace: 'data-register',
-                name: 'abcd',
-                latest: {
-                  configuration: ['123'],
-                },
-              },
-            ],
-          },
-        });
-
-      const handler = findDataRegisters(directoryMock, tokenProviderMock);
-      await handler(req as unknown as Request, res as unknown as Response, next);
-
-      expect(res.status).toHaveBeenCalledWith(200);
-      expect(res.send).toHaveBeenCalledWith([
-        {
-          name: 'abcd',
-          namespace: 'data-register',
-          description: '',
-          entries: ['123'],
-        },
-      ]);
-      expect(next).not.toHaveBeenCalled();
-    });
-
-    it('returns an empty array when configuration service returns no results', async () => {
-      const req = {
-        tenant: { id: tenantId },
-      };
-
-      const res = {
-        status: jest.fn().mockReturnThis(),
-        send: jest.fn(),
-      };
-
-      const next = jest.fn();
-
-      directoryMock.getServiceUrl.mockResolvedValueOnce(new URL('https://configuration-service/configuration/v2'));
-
-      axiosMock.get
-        .mockResolvedValueOnce({
-          data: {
-            latest: {
-              configuration: {},
-            },
-          },
-        })
-        .mockResolvedValueOnce({
-          data: {},
-        });
-
-      const handler = findDataRegisters(directoryMock, tokenProviderMock);
-      await handler(req as unknown as Request, res as unknown as Response, next);
-
-      expect(res.status).toHaveBeenCalledWith(200);
-      expect(res.send).toHaveBeenCalledWith([]);
-      expect(next).not.toHaveBeenCalled();
-    });
-
-    it('calls next when configuration service request fails', async () => {
-      const req = {
-        tenant: { id: tenantId },
-      };
-
-      const res = {
-        status: jest.fn().mockReturnThis(),
-        send: jest.fn(),
-      };
-
-      const next = jest.fn();
-      const error = new Error('configuration service error');
-
-      directoryMock.getServiceUrl.mockResolvedValueOnce(new URL('https://configuration-service/configuration/v2'));
-      axiosMock.get.mockRejectedValueOnce(error);
-
-      const handler = findDataRegisters(directoryMock, tokenProviderMock);
-      await handler(req as unknown as Request, res as unknown as Response, next);
-
-      expect(res.send).not.toHaveBeenCalled();
-      expect(next).toHaveBeenCalledWith(error);
-    });
-  });
-
-  describe('createDataRegister', () => {
-    it('can create handler', () => {
-      const handler = createDataRegister(directoryMock, tokenProviderMock);
-      expect(handler).toBeTruthy();
-    });
-
-    it('creates data register definition and configuration', async () => {
-      const user = {
-        tenantId,
-        id: 'tester',
-        roles: [FormServiceRoles.Admin],
-      };
-
-      const req = {
-        user,
-        tenant: { id: tenantId },
-        body: {
-          name: 'test-register',
-          description: 'Test register',
-          entries: ['one', 'two'],
-        },
-      };
-
-      const res = {
-        status: jest.fn().mockReturnThis(),
-        send: jest.fn(),
-      };
-
-      const next = jest.fn();
-
-      directoryMock.getServiceUrl.mockResolvedValueOnce(new URL('https://configuration-service/configuration/v2'));
-      axiosMock.patch
-        .mockResolvedValueOnce({
-          data: {
-            latest: {
-              revision: 1,
-              configuration: {},
-            },
-          },
-        })
-        .mockResolvedValueOnce({
-          data: {
-            latest: {
-              revision: 1,
-              configuration: ['one', 'two'],
-            },
-          },
-        });
-
-      const handler = createDataRegister(directoryMock, tokenProviderMock);
-      await handler(req as unknown as Request, res as unknown as Response, next);
-
-      expect(directoryMock.getServiceUrl.mock.calls[0][0].toString()).toBe('urn:ads:platform:configuration-service:v2');
-      expect(tokenProviderMock.getAccessToken).toHaveBeenCalled();
-      expect(axiosMock.patch).toHaveBeenCalledTimes(2);
-
-      expect(axiosMock.patch.mock.calls[0][0]).toContain('/configuration/platform/configuration-service');
-      expect(axiosMock.patch.mock.calls[0][1]).toEqual({
-        operation: 'UPDATE',
-        update: {
-          'data-register:test-register': {
-            configurationSchema: {
-              type: 'array',
-              items: {
-                type: 'string',
-              },
-            },
-            description: 'Test register',
-          },
-        },
-      });
-      expect(axiosMock.patch.mock.calls[0][2]).toEqual({
-        headers: { Authorization: 'Bearer token' },
-        params: { tenantId: tenantId.toString() },
-      });
-
-      expect(axiosMock.patch.mock.calls[1][0]).toContain('/configuration/data-register/test-register');
-      expect(axiosMock.patch.mock.calls[1][1]).toEqual({
-        operation: 'REPLACE',
-        configuration: ['one', 'two'],
-      });
-      expect(axiosMock.patch.mock.calls[1][2]).toEqual({
-        headers: { Authorization: 'Bearer token' },
-        params: { tenantId: tenantId.toString() },
-      });
-
-      expect(res.status).toHaveBeenCalledWith(201);
-      expect(res.send).toHaveBeenCalledWith({
-        namespace: 'data-register',
-        name: 'test-register',
-        description: 'Test register',
-        entries: ['one', 'two'],
-      });
-      expect(next).not.toHaveBeenCalled();
-    });
-
-    it('defaults optional description and entries when creating a data register', async () => {
-      const user = {
-        tenantId,
-        id: 'tester',
-        roles: [FormServiceRoles.Admin],
-      };
-
-      const req = {
-        user,
-        tenant: { id: tenantId },
-        body: {
-          name: 'empty-register',
-        },
-      };
-
-      const res = {
-        status: jest.fn().mockReturnThis(),
-        send: jest.fn(),
-      };
-
-      const next = jest.fn();
-
-      directoryMock.getServiceUrl.mockResolvedValueOnce(new URL('https://configuration-service/configuration/v2'));
-      axiosMock.patch
-        .mockResolvedValueOnce({
-          data: {
-            latest: {
-              revision: 1,
-              configuration: {},
-            },
-          },
-        })
-        .mockResolvedValueOnce({
-          data: {
-            latest: {
-              revision: 1,
-              configuration: [],
-            },
-          },
-        });
-
-      const handler = createDataRegister(directoryMock, tokenProviderMock);
-      await handler(req as unknown as Request, res as unknown as Response, next);
-
-      expect(axiosMock.patch.mock.calls[0][1]).toEqual({
-        operation: 'UPDATE',
-        update: {
-          'data-register:empty-register': {
-            configurationSchema: {
-              type: 'array',
-              items: {
-                type: 'string',
-              },
-            },
-            description: '',
-          },
-        },
-      });
-      expect(axiosMock.patch.mock.calls[1][1]).toEqual({
-        operation: 'REPLACE',
-        configuration: [],
-      });
-      expect(res.status).toHaveBeenCalledWith(201);
-      expect(res.send).toHaveBeenCalledWith({
-        namespace: 'data-register',
-        name: 'empty-register',
-        description: '',
-        entries: [],
-      });
-      expect(next).not.toHaveBeenCalled();
-    });
-
-    it('calls next with unauthorized error for non-admin users', async () => {
-      const user = {
-        tenantId,
-        id: 'tester',
-        roles: ['test-applicant'],
-      };
-
-      const req = {
-        user,
-        tenant: { id: tenantId },
-        body: {
-          name: 'test-register',
-        },
-      };
-
-      const res = {
-        status: jest.fn().mockReturnThis(),
-        send: jest.fn(),
-      };
-
-      const next = jest.fn();
-
-      const handler = createDataRegister(directoryMock, tokenProviderMock);
-      await handler(req as unknown as Request, res as unknown as Response, next);
-
-      expect(directoryMock.getServiceUrl).not.toHaveBeenCalled();
-      expect(axiosMock.patch).not.toHaveBeenCalled();
-      expect(res.send).not.toHaveBeenCalled();
-      expect(next.mock.calls[0][0]).toBeInstanceOf(UnauthorizedUserError);
     });
   });
 });
